Add system node usable for utils k3s cluster
Adds Terraform code to deploy a VM based on the k3s image snapshot
usable for a kubernetes cluster.

The utils cluster is intended to keep the Terraform state of the
deployment of the idunion kubernetes clusters. The kubernetes backend
allows state locking and gets encrypted at rest when k3s secret
encryption is enabled.

This is for now the easiest and secure way to save the Terraform state
remotely. Backup of the state needs still to be done, but should be
simple possible via volume snapshotting or copying the file.

For the current development phase of the idunion cluster, no utils
cluster will be used because of cost reasons. The TF state of the
development cluster will be stored locally which works fine with only
one developer.